Renamed server channel to rdpei
