+++ include/freerdp/server/rdpecam-enumerator.h
